Refactor (move CustomBottomAppBar to an own file)
