+++ lib/main.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'customBottomAppBar.dart';
+
 void main() {
   runApp(const MyApp());
 }
@@ -45,44 +47,6 @@ class HomePage extends StatelessWidget {
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
       bottomNavigationBar: const CustomBottomAppBar(),
-    );
-  }
-}
-
-class CustomBottomAppBar extends StatelessWidget {
-  const CustomBottomAppBar({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BottomAppBar(
-      shape: const CircularNotchedRectangle(),
-      color: Colors.blue,
-      child: IconTheme(
-        data: IconThemeData(color: Theme.of(context).colorScheme.onPrimary),
-        child: Row(
-          children: <Widget>[
-            const Spacer(
-              flex: 3,
-            ),
-            IconButton(
-              tooltip: 'Randomize',
-              icon: const Icon(Icons.shuffle),
-              onPressed: () {},
-            ),
-            const Spacer(
-              flex: 1,
-            ),
-            IconButton(
-              tooltip: 'Stop',
-              icon: const Icon(Icons.stop),
-              onPressed: () {},
-            ),
-            const Spacer(
-              flex: 3,
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
